User Infomation
Users can now add their view edit and  delete their information. All information is saved on the database under the customers table. Done CSS for accounts and customers page.

+++ Backend/database/migrations/2023_11_30_101155_create_products_table.php
@@ -15,8 +15,9 @@ return new class extends Migration
             $table->double('Price');
             $table->integer('StockQuantity')->default(0); // Removed quotes from default value
             $table->string('image')->nullable();
-            $table->unsignedBigInteger('CategoryID')->nullable(); // Changed to unsignedBigInteger
+            $table->unsignedInteger('CategoryID')->nullable(); // Changed to unsignedInteger
             $table->foreign('CategoryID')->references('CategoryID')->on('categories');
+
             $table->timestamps();
         });
     }
